Remove dead GVFSLock and GitStatusCache code
