+++ GVFS/GVFS.UnitTests/Mock/Common/MockPlatform.cs
@@ -151,11 +151,6 @@ namespace GVFS.UnitTests.Mock.Common
             throw new NotSupportedException();
         }
 
-        public override bool IsGitStatusCacheSupported()
-        {
-            return true;
-        }
-
         public override FileBasedLock CreateFileBasedLock(PhysicalFileSystem fileSystem, ITracer tracer, string lockPath)
         {
             return new MockFileBasedLock(fileSystem, tracer, lockPath);
