+++ GVFS/GVFS/CommandLine/StatusVerb.cs
@@ -34,7 +34,6 @@ namespace GVFS.CommandLine
                     this.Output.WriteLine("Cache Server: " + getStatusResponse.CacheServer);
                     this.Output.WriteLine("Local Cache: " + getStatusResponse.LocalCacheRoot);
                     this.Output.WriteLine("Mount status: " + getStatusResponse.MountStatus);
-                    this.Output.WriteLine("GVFS Lock: " + getStatusResponse.LockStatus);
                     this.Output.WriteLine("Background operations: " + getStatusResponse.BackgroundOperationCount);
                     this.Output.WriteLine("Disk layout version: " + getStatusResponse.DiskLayoutVersion);
                 }
